Refactor: CardTabMenu.tsx 생성
